+++ src/qt/winshutdownmonitor.cpp
@@ -5,7 +5,7 @@
 #include <qt/winshutdownmonitor.h>
 
 #if defined(Q_OS_WIN)
-#include <init.h>
+#include <shutdown.h>
 #include <util.h>
 
 #include <windows.h>
